+++ Assets/CommonCoreGame/RPGGame/RPG/InventoryModelTypes.cs
@@ -47,7 +47,7 @@ namespace CommonCore.RpgGame.Rpg
         WeaponPierceConsiderShields, WeaponPierceConsiderArmor, WeaponIgnoreShields, WeaponIgnoreArmor, WeaponNeverAlert, WeaponNeverBlockable, WeaponNoPain, WeaponAlwaysPain, WeaponIgnoreHitLocation, WeaponAlwaysExtremeDeath, WeaponNeverExtremeDeath,
 
         //melee-specific weapon flags
-        MeleeWeaponUsePreciseCasting, MeleeWeaponDelayCasting,
+        MeleeWeaponUsePreciseCasting, MeleeWeaponDelayCasting, MeleeWeaponAllowMultipleHits, MeleeWeaponHitNonDamageable, MeleeWeaponUseContactHitHack, MeleeWeaponDistinctMultipleHits,
 
         //dummy-specific weapon flags
         DummyWeaponUseViewModelRaiseLower,
@@ -398,9 +398,14 @@ namespace CommonCore.RpgGame.Rpg
         public readonly float EnergyCost;
         public readonly float DamageDelay;
 
+        public readonly float CastRadius;
+
+        public readonly string EnvironmentHitPuff;
+
         public MeleeWeaponItemModel(string name, float weight, float value, float maxCondition, int maxQuantity, bool hidden, bool essential, string[] flags, ItemScriptNode scripts,
             float damage, float damagePierce, float damageSpread, float damagePierceSpread,
             float reach, float rate, float energyCost, float damageDelay,
+            float castRadius, string environmentHitPuff,
             DamageType dType, DamageEffector? dEffector, WeaponSkillType skillType,
             string viewModel, string worldModel, string hitPuff, float lowerTime, float raiseTime) 
             : base(name, weight, value, maxCondition, maxQuantity, hidden, essential, flags, scripts, damage, damagePierce, damageSpread, damagePierceSpread, dType, dEffector, skillType, viewModel, worldModel, hitPuff, lowerTime, raiseTime)
@@ -409,6 +414,9 @@ namespace CommonCore.RpgGame.Rpg
             Rate = rate;
             EnergyCost = energyCost;
             DamageDelay = damageDelay;
+
+            CastRadius = castRadius;
+            EnvironmentHitPuff = environmentHitPuff;
         }
 
         public override DamageEffector Effector => DEffector ?? DamageEffector.Melee;