+++ CHANGELOG.md
@@ -184,3 +184,5 @@
 * Factions are now entirely case insensitive
 * Added support for glue-to-bone to ranged weapon viewmodel [Experimental]
 * Added recock handling to weapons [WIP]
+* Added multiple hit handling to melee weapons [WIP]
+* Added contact-hit hack to melee weapons [WIP]